third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/bounding-box-computation-3-expected.txt

This is a testharness.js-based test.
PASS Control test: Table width is 120px
PASS Control test: Table height is 120px
FAIL First (empty) table-row-group is 0px wide assert_equals: expected 0 but got 120
FAIL First (empty) table-row-group is 0px tall assert_equals: expected 0 but got 10
FAIL First (empty) table-row-group should be located at 10px left assert_equals: expected 10 but got 0
FAIL First (empty) table-row-group should be located at 10px top assert_equals: expected 10 but got 0
FAIL Second table-row-group is 100px wide assert_equals: expected 100 but got 120
FAIL Second table-row-group is 100px tall assert_equals: expected 100 but got 110
FAIL Second table-row-group should be located at 10px left assert_equals: expected 10 but got 0
PASS Second table-row-group should be located at 10px top
Harness: the test ran to completion.
